Refactor TeamsRepositoryImpl to remove UploadManager dependency
- Introduced TeamsUploadRepository and TeamsUploadRepositoryImpl to encapsulate team-specific upload orchestration (batching, correlation, and retries), extracting it from the generic UploadManager service.
- Replaced UploadManager dependency in TeamsRepositoryImpl with TeamsUploadRepository to resolve the architectural layering violation where a domain repository was improperly depending on a top-level service.
- Refactored UploadManager to delegate uploadTeams, uploadTeamActivities, and uploadResource to the newly instantiated TeamsUploadRepository rather than maintaining duplicate logic.
- Updated related unit tests in TeamsRepositoryImplTest and UploadManagerTest to correctly mock and verify against the injected TeamsUploadRepository interface.
