+++ app/src/main/java/org/ole/planet/myplanet/di/RepositoryModule.kt
@@ -159,6 +159,10 @@ abstract class RepositoryModule {
     @Binds
     @Singleton
     abstract fun bindUploadRepository(impl: UploadRepositoryImpl): UploadRepository
+    @Binds
+    @Singleton
+    abstract fun bindTeamsUploadRepository(impl: org.ole.planet.myplanet.repository.TeamsUploadRepositoryImpl): org.ole.planet.myplanet.repository.TeamsUploadRepository
+
 
     @Binds
     @Singleton
