+++ app/src/main/java/org/ole/planet/myplanet/repository/TeamsUploadRepository.kt
@@ -1,0 +1,9 @@
+package org.ole.planet.myplanet.repository
+
+import org.ole.planet.myplanet.callback.OnSuccessListener
+
+interface TeamsUploadRepository {
+    suspend fun uploadTeams()
+    suspend fun uploadTeamActivities()
+    suspend fun uploadResource(listener: OnSuccessListener?)
+}
