+++ app/src/main/java/org/ole/planet/myplanet/repository/UploadRepository.kt
@@ -14,8 +14,6 @@ interface UploadRepository {
     suspend fun putUpload(url: String, serializedData: JsonObject): Response<JsonObject>
     suspend fun fetchExistingDoc(url: String): Response<JsonObject>
     suspend fun uploadAttachment(file: java.io.File, destinationFormat: String, id: String, rev: String, name: String): Response<JsonObject>
-    suspend fun uploadTeams(teamsToUpload: List<org.ole.planet.myplanet.repository.TeamUploadData>): Response<com.google.gson.JsonArray>
-    suspend fun uploadTeamActivities(activitiesToUpload: com.google.gson.JsonArray): Response<com.google.gson.JsonArray>
 }
 
 data class UploadQueryContract<T : Any>(
